LayoutTests/platform/blackberry/fast/filesystem/simple-temporary-expected.txt

CONSOLE MESSAGE: line 24: ReferenceError: Can't find variable: webkitRequestFileSystem
FAIL: Timed out waiting for notifyDone to be called
requestFileSystem TEMPORARY test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
